+++ src/utils/settings.py
@@ -247,7 +247,7 @@ DESCRIPTION_FULL = "Automated All-in-One OS Command Injection Exploitation Tool"
 DESCRIPTION = "The command injection exploiter"
 AUTHOR  = "Anastasios Stasinopoulos"
 VERSION_NUM = "4.0"
-REVISION = "3"
+REVISION = "4"
 STABLE_RELEASE = False
 VERSION = "v"
 if STABLE_RELEASE:
@@ -1268,6 +1268,9 @@ SKIP_VULNERABLE_HOST = None
 
 # Skipped crawled hrefs
 HREF_SKIPPED = []
+
+# Abort on (problematic) HTTP error code (e.g. 401).
+ABORT_CODE = []
 
 # Default crawling depth
 DEFAULT_CRAWLING_DEPTH = 1